Feat: PR 리뷰 후속 커밋
Validation과 전역 예외 처리 구현
AuthorityStatusConverter 구현
ErrorCode 상황에 맞는 HttpStatus로 수정
