+++ api/src/main/java/com/zerobase/commerce/api/exception/CustomExceptionHandler.java
@@ -1,9 +1,14 @@
 package com.zerobase.commerce.api.exception;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.support.DefaultMessageSourceResolvable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.stream.Collectors;
 
 @RestControllerAdvice
 @Slf4j
@@ -11,6 +16,12 @@ public class CustomExceptionHandler {
     @ExceptionHandler(CustomException.class)
     public ResponseEntity<?> customExceptionHandler(CustomException e) {
         return new ErrorResponse(e).toResponseEntity();
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<?> validationException(MethodArgumentNotValidException e) {
+        var errorMessages = e.getBindingResult().getAllErrors().stream().map(DefaultMessageSourceResolvable::getDefaultMessage).toList();
+        return new ErrorResponse(new CustomException(ErrorCode.VALIDATION_ERROR)).toResponseEntity(String.join(", ", errorMessages) + ".");
     }
 
     @ExceptionHandler(RuntimeException.class)
