+++ api/src/main/java/com/zerobase/commerce/api/user/dto/UpdateUserInfo.java
@@ -1,5 +1,6 @@
 package com.zerobase.commerce.api.user.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,6 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateUserInfo {
+    @NotBlank
     private String password;
     private String newPassword;
 }
