+++ api/src/test/java/com/zerobase/commerce/api/user/controller/UserControllerTest.java
@@ -2,8 +2,7 @@ package com.zerobase.commerce.api.user.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zerobase.commerce.api.security.TokenProvider;
-import com.zerobase.commerce.api.user.dto.UpdateUserInfo;
-import com.zerobase.commerce.api.user.dto.UserDto;
+import com.zerobase.commerce.api.user.dto.*;
 import com.zerobase.commerce.api.user.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -38,6 +37,49 @@ class UserControllerTest {
 
     @MockBean
     private TokenProvider tokenProvider;
+
+    @Test
+    @DisplayName("회원가입 성공")
+    void signUpSuccess() throws Exception {
+        var request = SignUpDto.Request.builder()
+                .id("root1234")
+                .password("qwer1234")
+                .build();
+
+        given(userService.signUp(any())).willReturn(
+                SignUpDto.Response.builder()
+                        .id("root1234")
+                        .password("qwer1234")
+                        .roles(Arrays.stream(new String[]{"ROLE_MEMBER"}).collect(Collectors.toSet()))
+                        .build());
+
+        var result = mockMvc.perform(post("/auth/sign-up")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(request)));
+
+        result.andExpect(status().isOk())
+                .andExpect(jsonPath("$.id").value("root1234"))
+                .andExpect(jsonPath("$.password").value("qwer1234"))
+                .andExpect(jsonPath("$.roles[0]").value(ROLE_MEMBER.name()))
+                .andDo(print());
+    }
+
+    @Test
+    @DisplayName("로그인 성공")
+    void signInSuccess() throws Exception {
+        var request = SignInDto.Request.builder()
+                .id("root1234")
+                .password("qwer1234")
+                .build();
+
+        given(userService.signIn(any())).willReturn("token");
+
+        var result = mockMvc.perform(post("/auth/sign-in")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(request)));
+
+        result.andExpect(status().isOk());
+    }
 
     @Test
     @DisplayName("유저 정보 가져오기 성공")
