+++ database/src/main/java/com/zerobase/commerce/database/domain/converter/AuthorityStatusConverter.java
@@ -1,0 +1,23 @@
+package com.zerobase.commerce.database.domain.converter;
+
+import com.zerobase.commerce.database.constant.AuthorityStatus;
+import jakarta.persistence.AttributeConverter;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+public class AuthorityStatusConverter implements AttributeConverter<Set<AuthorityStatus>, String> {
+    private final String SEPARATOR = ",";
+
+    @Override
+    public String convertToDatabaseColumn(Set<AuthorityStatus> attribute) {
+        return String.join(SEPARATOR, attribute.stream().map(AuthorityStatus::toString).collect(Collectors.toSet()));
+    }
+
+    @Override
+    public Set<AuthorityStatus> convertToEntityAttribute(String dbData) {
+        return dbData.isEmpty() ? new HashSet<>() : Arrays.stream(dbData.split(SEPARATOR)).map(AuthorityStatus::valueOf).collect(Collectors.toSet());
+    }
+}
